Splitting Agents Logic in multiple files.
