+++ playground/react/src/App.jsx
@@ -1,5 +1,5 @@
 import AgentsBoardDebugger from './AgentsBoardDebugger';
-import team from './teams/openai/productSpecsTeam';
+import team from './teams/openai/sportsNewsTeam';
 
 function App() {
   return (
