+++ playground/react/src/teams/openai/productSpecsTeam.js
@@ -22,26 +22,26 @@ const requirementsAnalyst = new Agent({
     goal: 'Outline core functionalities and objectives for new features based on the founder’s input.',
     background: 'Business Analysis',
     tools: [search],
-    // type: 'ReAct',
+    type: 'ReAct',
 });
 
-const technicalWriter = new Agent({
-    name: 'Lucas',
-    role: 'Technical Writer',
-    goal: 'Convert functional outlines into detailed technical specifications.',
-    background: 'Technical Writing',
-    tools: [search],
-    // type: 'ReAct',
-});
+// const technicalWriter = new Agent({
+//     name: 'Lucas',
+//     role: 'Technical Writer',
+//     goal: 'Convert functional outlines into detailed technical specifications.',
+//     background: 'Technical Writing',
+//     tools: [search],
+//     // type: 'ReAct',
+// });
 
-const validator = new Agent({
-    name: 'Mia',
-    role: 'Validator',
-    goal: 'Ensure the specifications are accurate and complete.',
-    background: 'Quality Assurance',
-    tools: [search],
-    // type: 'ReAct',
-});
+// const validator = new Agent({
+//     name: 'Mia',
+//     role: 'Validator',
+//     goal: 'Ensure the specifications are accurate and complete.',
+//     background: 'Quality Assurance',
+//     tools: [search],
+//     // type: 'ReAct',
+// });
 
 // Define tasks
 const analysisTask = new Task({
@@ -50,26 +50,32 @@ const analysisTask = new Task({
     agent: requirementsAnalyst
 });
 
-const writingTask = new Task({
-    description: `Create detailed technical specifications based on the functional outline provided. Include user stories, system requirements, and acceptance criteria.`,
-    expectedOutput: 'A detailed technical specifications document.',
-    isDeliverable: true,
-    agent: technicalWriter
-});
+// const writingTask = new Task({
+//     description: `Create detailed technical specifications based on the functional outline provided. Include user stories, system requirements, and acceptance criteria.`,
+//     expectedOutput: 'A detailed technical specifications document.',
+//     isDeliverable: true,
+//     agent: technicalWriter
+// });
 
-const validationTask = new Task({
-    description: `Review the technical specifications to ensure they match the founder's vision and that are technically feasible.`,
-    expectedOutput: 'A validated technical specifications document ready for development. Must be in Markdown format.',
-    agent: validator
-});
+// const validationTask = new Task({
+//     description: `Review the technical specifications to ensure they match the founder's vision and that are technically feasible.`,
+//     expectedOutput: 'A validated technical specifications document ready for development. Must be in Markdown format.',
+//     agent: validator
+// });
 
 // Create a team
+// const team = new Team({
+//     name: 'Product Specs Team',
+//     agents: [requirementsAnalyst, technicalWriter, validator],
+//     tasks: [analysisTask, writingTask, validationTask],
+//     inputs: { founderIdea: 'I want to add a Referral program to our SAAS platform.' },  // Initial input for the first task
+// });
+
 const team = new Team({
     name: 'Product Specs Team',
-    agents: [requirementsAnalyst, technicalWriter, validator],
-    tasks: [analysisTask, writingTask, validationTask],
+    agents: [requirementsAnalyst],
+    tasks: [analysisTask],
     inputs: { founderIdea: 'I want to add a Referral program to our SAAS platform.' },  // Initial input for the first task
-    // env: { OPENAI_API_KEY: import.meta.env.VITE_OPENAI_API_KEY }
 });
 
 export default team;
